requirements.txt
asgiref==3.11.1
attrs==26.1.0
coverage==7.13.5
Django==6.0.4
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
factory_boy==3.3.3
Faker==40.15.0
gunicorn==25.3.0
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
packaging==26.1
pluggy==1.6.0
psycopg2-binary==2.9.12
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.3
pytest-django==4.12.0
python-decouple==3.8
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
